+++ extensions/zeabur/CHANGELOG.md
@@ -1,3 +1,9 @@
-# Zeabur Docs Changelog
+# Zeabur Changelog
 
-## [Initial Version] - 2024-09-12
+## [Feature] - {PR_MERGE_DATE}
+
+- Added the deploy project command to Zeabur.
+
+## [Feature] - 2024-09-12
+
+- Added search Zeabur documentation command.